Separate Podcast feature into a new module and move GraphQL Schema into Core module
