+++ app/build.gradle
@@ -1,7 +1,6 @@
 plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
-    id 'com.apollographql.apollo3' version "$apollo_version"
 }
 
 android {
@@ -49,6 +48,10 @@ android {
 
 dependencies {
 
+    //Modules
+    implementation project(':podcast_details')
+    implementation project(':core')
+
     def composeBom = platform('androidx.compose:compose-bom:2023.01.00')
     implementation composeBom
     androidTestImplementation composeBom
@@ -67,17 +70,10 @@ dependencies {
     // Navigation Compose
     implementation "androidx.navigation:navigation-compose:2.6.0-alpha04"
 
-    // Coroutines
-    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.4"
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.4'
-
     // Koin: Dependency Injection
     implementation "io.insert-koin:koin-core:$koin_version"
     implementation "io.insert-koin:koin-androidx-workmanager:$koin_android_version"
     implementation "io.insert-koin:koin-androidx-compose:$koin_android_compose_version"
-
-    // Apollo GraphQL Impl
-    implementation "com.apollographql.apollo3:apollo-runtime:$apollo_version"
 
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
@@ -85,10 +81,4 @@ dependencies {
     androidTestImplementation "androidx.compose.ui:ui-test-junit4"
     debugImplementation "androidx.compose.ui:ui-tooling"
     debugImplementation "androidx.compose.ui:ui-test-manifest"
-}
-
-apollo {
-    service("service") {
-        packageName.set("com.example.podcastapp")
-    }
 }
