+++ app/src/main/java/com/example/podcastapp/MainActivity.kt
@@ -10,9 +10,16 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.lifecycle.lifecycleScope
+import com.example.podcast_details.data.repositories.IPodcastRepository
 import com.example.podcastapp.ui.theme.PodcastAppTheme
+import org.koin.android.ext.android.inject
+import org.koin.java.KoinJavaComponent.inject
 
 class MainActivity : ComponentActivity() {
+
+    val repository by inject<IPodcastRepository>()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
@@ -25,6 +32,10 @@ class MainActivity : ComponentActivity() {
                     Greeting("Android")
                 }
             }
+        }
+
+        lifecycleScope.launchWhenResumed {
+            repository.getPodcastDetails("This American Life")
         }
     }
 }
