--- app/src/main/java/com/example/podcastapp/di/RepositoriesModule.kt
@@ -1,13 +1,0 @@
-package com.example.podcastapp.di
-
-import com.example.podcastapp.data.repositories.IPodcastRepository
-import com.example.podcastapp.data.repositories.PodcastRepository
-import org.koin.dsl.module
-
-val repositoryModule = module {
-    single<IPodcastRepository> {
-        PodcastRepository(
-            get()
-        )
-    }
-}
