+++ core/src/main/graphql/com/example/podcastapp_core/SearchForTerm.graphql
