+++ core/src/main/graphql/com/example/podcastapp_core/schema.sdl
